Add U.S. Weather Data (Source: NOAA)
Data obtained from the U.S. National Oceanic and Atmospheric Administration's (NOAA) tracks characteristics of major storms and weather events in the United States, including when and where they occur, as well as estimates of any fatalities, injuries, and property damage.

Note: data is in the form of a comma-separated-value file compressed via the bzip2 algorithm.